refactor create_billable_for_sms to be at the successful endpoint of sending all and receiving all sms messages
